move `compute_goal` and `evaluate_x` methods to inner module
